% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_profiling.R
\name{categ_analysis}
\alias{categ_analysis}
\title{Profiling analysis of categorical vs. target variable}
\usage{
categ_analysis(data, str_input, str_target)
}
\arguments{
\item{data}{input data containing the variable to describe}

\item{str_input}{string input variable (if empty, it runs for all categorical variable), it can take a single character value or a character vector.}

\item{str_target}{string target variable. Binary or two class is only supported by now.}
}
\value{
if str_input has 1 variable, it retrurns a data frame indicating all the metrics, otherwise prints in console all variable results.
}
\description{
Retrieves a complete summary of the grouped input variable against the target variable. Type of target variable must be binary for now. A positive case will be the less representative one.
It returns the total positive cases (sum_{STR_TARGET}); pecentage of total positive cases (perc_{STR_TARGET}) that fell in that category (this column sums 1); likelihood or mean of positive cases
 (mean_{STR_TARGET}) measured by the total positive cases over total cases in that category;
quantity of rows of that category (q_rows) and in percentage (perc_rows) -this column sums 1.
}
\examples{
categ_analysis(country, "country", "has_flu")
}

