% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numbers.R
\name{correlation_table}
\alias{correlation_table}
\title{Correlation analyisis against target variable}
\usage{
correlation_table(data, str_target)
}
\arguments{
\item{data}{data frame}

\item{str_target}{string variable to predict}
}
\value{
Correlation index for all data input variable
}
\description{
Obtain correlation table of all variables that belongs to data against target variable. Only numeric variables are analyzed.
}
\examples{
correlation_table(data=heart_disease, str_target="has_heart_disease")
}

