% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cross_plot.R
\name{cross_iterator}
\alias{cross_iterator}
\title{Generate cross_plot for all variables}
\usage{
cross_iterator(str_vars, data, str_target, path_out, auto_binning)
}
\arguments{
\item{str_vars}{string vector containing variable names of 'data' param.}

\item{data}{data frame source}

\item{str_target}{string of the variable to predict}

\item{path_out}{path directory, if it has a value the plot is saved}

\item{auto_binning}{indicates the automatic binning of str_input variable based on equal frequency (function 'equal_freq'), default value=TRUE}
}
\description{
Internal function used in massive_cross_plot to generate as many cross_plot as elements are in str_vars.
}

