% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cross_plot.R
\name{cross_plot}
\alias{cross_plot}
\title{Cross-plotting input variable vs. target variable}
\usage{
cross_plot(data, str_input, str_target, path_out, auto_binning)
}
\arguments{
\item{data}{data frame source}

\item{str_input}{string input variable}

\item{str_target}{string of the variable to predict}

\item{path_out}{path directory, if it has a value the plot is saved}

\item{auto_binning}{indicates the automatic binning of str_input variable based on equal frequency (function 'equal_freq'), default value=TRUE}
}
\value{
cross plot
}
\description{
The cross_plot shows how the input variable is correlated with the target variable, getting the likelihood rates for each input's bin/bucket .
}
\examples{
## Example 1:
cross_plot(data=heart_disease, str_input="chest_pain", str_target="has_heart_disease")

## Example 2:
cross_plot(data=heart_disease, str_input="age", str_target="has_heart_disease")
}

