% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_country}
\alias{data_country}
\title{People with flu data}
\format{A data frame with 910 rows and 3 variables}
\usage{
data_country
}
\description{
Each row represents a person from different countries indicating if he or she has or not flu.
Colmuns
person: unique id
country: country of the person, 70 different countries
has_flu: character variable with values "yes" or "no" indicating if the person has flu
}
\keyword{datasets}
