% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_golf}
\alias{data_golf}
\title{Play golf}
\format{
A data frame with 14 rows and 3 variables
}
\usage{
data_golf
}
\description{
This well known small data frame containst 14 cases indicating wheter or not play golf based on wheather conditions. Target variable: 'play_golf.'
}
\keyword{datasets}
