% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecting_best_variables.R
\name{df_recover_cuts}
\alias{df_recover_cuts}
\title{xx}
\usage{
df_recover_cuts(data, data_cuts, stringsAsFactors = F)
}
\arguments{
\item{data}{data frame source}

\item{data_cuts}{xx}

\item{stringsAsFactors}{xx}
}
\value{
xxx
}
\description{
xxx
}
\examples{
\dontrun{
# Reducing quantity of countries based on has_flu variable
xxx
}
}
