% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecting_best_variables.R
\name{discretize_df}
\alias{discretize_df}
\title{xx}
\usage{
discretize_df(data, data_bins, stringsAsFactors = F)
}
\arguments{
\item{data}{data frame source}

\item{stringsAsFactors}{xx}

\item{data_cuts}{xx}
}
\value{
xxx
}
\description{
xxx
}
\examples{
\dontrun{
# Reducing quantity of countries based on has_flu variable
xxx
}
}
