% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecting_best_variables.R
\name{discretize_get_bins}
\alias{discretize_get_bins}
\title{Get the cuts, or bin threshold, to discretize a variable.}
\usage{
discretize_get_bins(data, input = NULL, n_bins = 5, stringsAsFactors = F)
}
\arguments{
\item{data}{Data frame source}

\item{input}{Vector of string containing all the variables that will be processed.
If empty it will run for all numerical variables. NAs values are automatically handled converting
them into another category.}

\item{n_bins}{The number of bins (or segments) that each variable will have.
If it is needed a different number of bins per variable, then the function must be called more than once.}

\item{stringsAsFactors}{Boolean variable which indicates if the discretization result is character or factor.
When TRUE, the segments are ordered. TRUE by default.}
}
\value{
Data frame containing the thresholds or cuts to bin every variable
}
\description{
It takes a data frame
}
\examples{
\dontrun{
# Getting the bins thresholds for each. If input is missing, will run for all numerical variables.
d_bins=discretize_get_bins(data=heart_disease, input=c("resting_blood_pressure", "oldpeak"), n_bins=5)

# Now it can be applied on the same data frame, or in a new one (for example in a predictive model that change data over time)
heart_disease_discretized=discretize_df(data=heart_disease, data_bins=d_bins, stringsAsFactors=T)

}
}
