% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lares_model_plots.R
\name{dist2d}
\alias{dist2d}
\title{Distance from specific point to line}
\usage{
dist2d(a, b = c(0, 0), c = c(1, 1))
}
\arguments{
\item{a}{Vector. Coordinates of the point from which we want to
measure the distance}

\item{b}{Vector. Coordinates of 1st point over the line}

\item{c}{Vector. Coordinates of 2st point over the line}
}
\description{
This function lets the user calculate the mathematical linear distance
Between a specific point and a line (given geometrical 3 points)
}
\seealso{
Other Calculus: \code{\link{ROC}}, \code{\link{errors}},
  \code{\link{mae}}, \code{\link{mape}}, \code{\link{mse}},
  \code{\link{rmse}}, \code{\link{rsqa}}, \code{\link{rsq}}
}
\author{
Bernardo Lares
}
\concept{Calculus}
