% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{entropy_2}
\alias{entropy_2}
\title{Calculate entropy between two variables}
\usage{
entropy_2(x, y)
}
\arguments{
\item{x}{input numeric vector}

\item{y}{input numeric vector}
}
\description{
It calculates the entropy between two categorical variables using log2.
This log2 is mentioned in most of the Claude Shannon bibliography.
Input vectors, 'x' and 'y' can be numeric or character. REVISAR QUE CALCULO HAGO ACA
}
\examples{
\dontrun{
entropy_2(heart_disease$gender, heart_disease$chest_pain)
}
}
