% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_lib.R
\name{export_plot}
\alias{export_plot}
\title{Export plot to jpeg file}
\usage{
export_plot(object_plot, path_out, file_name)
}
\arguments{
\item{object_plot}{Object plot to export (like ggplot2)}

\item{path_out}{path directory to export the output, if it has a value the plot is saved,
if the directory doesn't existis it will try to create it. To save in current directory path must be dot: "."}

\item{file_name}{output file name}
}
\value{
none
}
\description{
Export 'object_plot' to jpeg file under the name 'file_name' in the directory 'path_out'
}
