% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_lib.R
\name{fibonacci}
\alias{fibonacci}
\title{Fibonacci series}
\usage{
fibonacci(length, remove_first = F)
}
\arguments{
\item{length}{data frame}

\item{remove_first}{removes the first value of the series, because first 2 elements are the same (number=1). False by default.}
}
\value{
vector
}
\description{
It retrieves a vector containing the first N numbers specified in 'length' parameter of the Fibonacci series.
}
\examples{
# Get the first 4 elements of Fibonacci series
fibonacci(4)
}

