% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numbers.R
\name{freq}
\alias{freq}
\title{Frequency table for categorical variables}
\usage{
freq(data, str_input)
}
\arguments{
\item{data}{input data containing the variable to describe}

\item{str_input}{string variable name}
}
\value{
vector with the values scaled into the 0 to 1 range
}
\description{
Retrieves the frequency and percentage for str_input
}
\examples{
freq(mtcars, 'gear')
}

