% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funModeling.R
\docType{package}
\name{funModeling-package}
\alias{funModeling}
\alias{funModeling-package}
\title{funModeling: Exploratory data analysis, data preparation and model performance}
\description{
funModeling is intimately related to the Data Science Live Book -Open Source- (2017) in the sense that most
of its functionality is used to explain different topics addressed by the book.
}
\details{
To start using funModeling you can start by the vignette:
`browseVignettes(package = "funModeling")`

Or you can read the Data Science Live Book, fully accessible at: \url{https://livebook.datascienceheroes.com}
}
\seealso{
Useful links:
\itemize{
  \item \url{livebook.datascienceheroes.com}
  \item Report bugs at \url{https://github.com/pablo14/funModeling/issues}
}

}
\author{
\strong{Maintainer}: Pablo Casas \email{pcasas.biz@gmail.com}

}
