% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{gain_ratio}
\alias{gain_ratio}
\title{Gain ratio}
\usage{
gain_ratio(input, target)
}
\arguments{
\item{input}{numeric/character vector}

\item{target}{numeric/character vector}
}
\value{
gain ratio
}
\description{
Computes the information gain between an 'input' and 'target' variable (using log2). Similar to information gain but less sensitive to high cardinality variables.
}
\examples{
\dontrun{
gain_ratio(input=data_golf$outlook, target=data_golf$play_golf)
}
}
