% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numbers.R
\name{get_cor_matrix}
\alias{get_cor_matrix}
\title{Plot}
\usage{
get_cor_matrix(data, target)
}
\arguments{
\item{data}{data frame}
}
\value{
Metrics data frame
}
\description{
For each variable it returns: Quantity and percentage of zeros (q_zeros and p_zeros respectevly). Same metrics for NA values (q_NA and p_na). Last two columns indicates data type and quantity of unique values.
This function print and return the results.
}
\examples{
df_status(heart_disease)
}

