% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lares_model_plots.R
\name{gg_text_customs}
\alias{gg_text_customs}
\title{Custom colours to use in ggplot as scale_color_manual on texts}
\usage{
gg_text_customs()
}
\description{
This function lets the user use pre-defined default colours
}
\seealso{
Other Auxiliary: \code{\link{gg_colour_customs}},
  \code{\link{gg_fill_customs}}, \code{\link{lares_pal}},
  \code{\link{plot_palette}}
}
\author{
Bernardo Lares
}
\concept{Auxiliary}
