% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heart_disease}
\alias{heart_disease}
\title{Heart Disease Data}
\format{
A data frame with 303 rows and 16 variables:
\itemize{

   \url{https://archive.ics.uci.edu/ml/datasets/Heart+Disease}
}
}
\usage{
heart_disease
}
\description{
There are variables related to patient clinic trial. The variable to predict is `has_heart_disease`.
}
\keyword{datasets}
