% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{infor_magic}
\alias{infor_magic}
\title{Computes several information theory metrics between two vectors}
\usage{
infor_magic(input, target)
}
\arguments{
\item{input}{vector to be evaluated against the variable defined in 'target' parameter}

\item{target}{vector containing the output variable.}
}
\value{
Matrix of 1 row and 4 columns, where each column represent the mentioned metrics
}
\description{
It retrieves the same as \code{\link{var_rank_info}} but receiving two vectors.
Metrics are: entropy (en), mutual information (mi), information gain (ig) and gain ratio (gr).
}
\examples{
\dontrun{
infor_magic(data_golf$outlook, data_golf$play_golf)
}
}
