% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lares_model_plots.R
\name{lares_pal}
\alias{lares_pal}
\title{Personal Colours Palette}
\usage{
lares_pal()
}
\description{
This function plots a list of colours on a specific palette
}
\seealso{
Other Auxiliary: \code{\link{gg_colour_customs}},
  \code{\link{gg_fill_customs}},
  \code{\link{gg_text_customs}}, \code{\link{plot_palette}}
}
\author{
Bernardo Lares
}
\concept{Auxiliary}
