% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models_lib.R
\name{lift_table}
\alias{lift_table}
\title{Generates lift or gain performance table}
\usage{
lift_table(data, str_score, str_target, q_segments)
}
\arguments{
\item{data}{input data source}

\item{str_score}{the variable which contains the score number}

\item{str_target}{target variable}

\item{q_segments}{quantity of segments to split str_score_var, valid values: 5, 10 or 20}
}
\value{
lift/gain table, column: gain implies how much positive cases are catched if the cut point to define the positive class is set to the column "Score Point"
}
\description{
It retrieves the cumulative positive rate when score is divided in (i.e.) 10 segments.
}
\examples{
fit_glm=glm(has_heart_disease ~ age + oldpeak, data=heart_disease, family = binomial)
heart_disease$score=predict(fit_glm, newdata=heart_disease, type='response')
lift_table(data=heart_disease,str_score='score',str_target='has_heart_disease')
}

