% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lares_model_plots.R
\name{mplot_lineal}
\alias{mplot_lineal}
\title{Linear Regression Results Plot}
\usage{
mplot_lineal(tag, score, subtitle = NA, model_name = NA,
  save = FALSE, subdir = NA, file_name = "viz_lineal.png")
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}

\item{subtitle}{Character. Subitle to show in plot}

\item{model_name}{Character. Model's name}

\item{save}{Boolean. Save output plot into working directory}

\item{subdir}{Character. Sub directory on which you wish to save the plot}

\item{file_name}{Character. File name as you wish to save the plot}
}
\description{
This function plots a Linear Regression Result
}
\seealso{
Other Machine Learning: \code{\link{ROC}},
  \code{\link{mplot_cuts_error}}, \code{\link{mplot_cuts}},
  \code{\link{mplot_density}}, \code{\link{mplot_full}},
  \code{\link{mplot_metrics}}, \code{\link{mplot_roc}},
  \code{\link{mplot_splits}}

Other Visualization: \code{\link{mplot_cuts_error}},
  \code{\link{mplot_cuts}}, \code{\link{mplot_density}},
  \code{\link{mplot_full}}, \code{\link{mplot_metrics}},
  \code{\link{mplot_roc}}, \code{\link{mplot_splits}},
  \code{\link{theme_lares2}}, \code{\link{theme_lares}}
}
\author{
Bernardo Lares
}
\concept{Machine Learning}
\concept{Visualization}
