% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numbers.R
\name{range01}
\alias{range01}
\title{Sigmoid function
#}
\usage{
range01(x)
}
\arguments{
\item{x}{numeric input vector
#}

\item{a}{constant to multiply 'x', default=1
#}

\item{x}{numeric input vector}
}
\value{
vector transformed with sigmoid
#

vector ranged into 0-1
}
\description{
Sigmoid function, also known as logistic or s-shaped
#
}
\details{
#
}
\examples{
#sigmoid()
#
}

