% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{reduce_cardinality}
\alias{reduce_cardinality}
\title{Reduce cardinality in categorical variable}
\usage{
reduce_cardinality(data, str_input, str_target, n_groups, seed = 999)
}
\arguments{
\item{data}{data frame source}

\item{str_input}{categorical variable indicating}

\item{str_target}{string of the variable to optimize the re-grouping}

\item{n_groups}{number of groups for the new category based on str_input, normally between ~3 and ~10.}

\item{seed}{optional, random number used internally for the k-means, changing this value will change the model}
}
\value{
A list containing 3 elements: recateg_results which contains the description of the target variable with the new groups;
df_equivalence is a data frame containing the str_input category and the new category; fit_cluster which is the cluster model used to do the re-grouping
}
\description{
Reduce the cardinality of an input variable based on a target -binary for now- variable based on attribitues of
accuracy and representativity, for both input and target variable. It uses a cluster model to create the new groups
}
\examples{
# Reducing quantity of countries based on has_flu variable
reduce_cardinallity(data=data_country, str_input='country', str_target="has_flu")
}

