% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lares_model_plots.R
\name{ROC}
\alias{ROC}
\title{ROC Curves}
\usage{
ROC(tag, score, multis = NA)
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}

\item{multis}{Data.frame. Containing columns with each category score
(only used when more than 2 categories coexist)}
}
\description{
This function calculates ROC Curves and AUC values with 95\% confidence
range. It also works for multi-categorical models.
}
\seealso{
Other Machine Learning: \code{\link{mplot_cuts_error}},
  \code{\link{mplot_cuts}}, \code{\link{mplot_density}},
  \code{\link{mplot_full}}, \code{\link{mplot_lineal}},
  \code{\link{mplot_metrics}}, \code{\link{mplot_roc}},
  \code{\link{mplot_splits}}

Other Calculus: \code{\link{dist2d}}, \code{\link{errors}},
  \code{\link{mae}}, \code{\link{mape}}, \code{\link{mse}},
  \code{\link{rmse}}, \code{\link{rsqa}}, \code{\link{rsq}}
}
\author{
Bernardo Lares
}
\concept{Calculus}
\concept{Machine Learning}
