% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lares_model_plots.R
\name{rsq}
\alias{rsq}
\title{R Squared}
\usage{
rsq(tag, score)
}
\arguments{
\item{tag}{Vector. Real known label}

\item{score}{Vector. Predicted value or model's result}
}
\description{
This function lets the user calculate r squared
}
\seealso{
Other Calculus: \code{\link{ROC}}, \code{\link{dist2d}},
  \code{\link{errors}}, \code{\link{mae}},
  \code{\link{mape}}, \code{\link{mse}},
  \code{\link{rmse}}, \code{\link{rsqa}}
}
\author{
Bernardo Lares
}
\concept{Calculus}
