% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numbers.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function}
\usage{
sigmoid(x, a = 1)
}
\arguments{
\item{x}{numeric input vector}

\item{a}{constant to multiply 'x', default=1}
}
\value{
vector transformed with sigmoid
}
\description{
Sigmoid function, also known as logistic or s-shaped
}
\examples{
sigmoid()
}

