% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_integrity.R
\name{status}
\alias{status}
\title{Get a summary for the given data frame (o vector).}
\usage{
status(data)
}
\arguments{
\item{data}{data frame, tibble or a single vector}
}
\value{
Tibble with metrics
}
\description{
For each variable it returns: Quantity and percentage of zeros (q_zeros and p_zeros respectevly). Same metrics for NA values (q_NA/p_na), and infinite values (q_inf/p_inf). Last two columns indicates data type and quantity of unique values.
'status' function is the evolution of 'df_status'. Main change is to have the decimal points as it is, except in percentage. For example now p_na=0.04 means 4% in df_status.
This time it's easier to embbed in a data process flow and to take actions based on this number.
}
\examples{
status(heart_disease)
}
