% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_data_analysis.R
\name{v_compare}
\alias{v_compare}
\title{Compare two vectors}
\usage{
v_compare(vector_x, vector_y)
}
\arguments{
\item{vector_x}{1st vector to compare}

\item{vector_y}{2nd vector to compare}
}
\value{
Correlation index for all data input variable
}
\description{
Obtaing coincident and not coincident elements between two vectors.
}
\examples{
v1=c("height","weight","age")
v2=c("height","weight","location","q_visits")
res=v_compare(vector_x=v1, vector_y=v2)
# Print the keys that didn't match
res
# Accessing the keys not present in
}
