% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{var_rank_info}
\alias{var_rank_info}
\title{Importance variable ranking based on information theory}
\usage{
var_rank_info(data, target)
}
\arguments{
\item{data}{input data frame, all the variables will be evaluated against the variable defined in 'target' parameter}

\item{target}{string variable name containing the output variable.}
}
\value{
data frame ordered by gain ratio metric
}
\description{
Retrieves a data frame containing several metrics related to information theory.
Metrics are: entropy (en), mutual information (mi), information gain (ig) and gain ratio (gr).
}
\examples{
\dontrun{
var_rank_info(data_golf, "play_golf")
}
}
