% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_profiling.R
\name{categ_analysis}
\alias{categ_analysis}
\title{Profiling analysis of categorical vs. target variable}
\usage{
categ_analysis(data, input, target)
}
\arguments{
\item{data}{input data containing the variable to describe}

\item{input}{string input variable (if empty, it runs for all categorical variable), it can take a single character value or a character vector.}

\item{target}{string target variable. Binary or two class is only supported by now.}
}
\value{
if str_input has 1 variable, it retrurns a data frame indicating all the metrics, otherwise prints in console all variable results.
}
\examples{
categ_analysis(country, "country", "has_flu")
}

