% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numbers.R
\name{compare_df}
\alias{compare_df}
\title{Compare two vectors of keys}
\usage{
compare_df(key_x, key_y)
}
\arguments{
\item{data}{data frame}

\item{str_target}{string variable to predict}
}
\value{
Correlation index for all data input variable
}
\description{
Obtain correlation table of all variables that belongs to data against target variable
}
\examples{
v1=c(1,2,4)
v2=c(1,2,5,6)
res=compare_df(key_x=v1, key_y=v2)
# Print the keys that didn't match
res
# Accessing the keys not present in
}

