% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_data_analysis.R
\name{correlation_table}
\alias{correlation_table}
\title{Get correlation against target variable}
\usage{
correlation_table(data, target)
}
\arguments{
\item{data}{data frame}

\item{target}{string variable to predict}
}
\value{
Correlation index for all data input variable
}
\description{
Obtain correlation table for all variables against target variable. Only numeric variables are analyzed (factor/character are skippted automatically).
}
\examples{
correlation_table(data=heart_disease, target="has_heart_disease")
}
