% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_lib.R
\name{df_status}
\alias{df_status}
\title{Get a summary for the given data frame.}
\usage{
df_status(data, print_results)
}
\arguments{
\item{data}{data frame}

\item{print_results}{if FALSE then there is not a print in the console, TRUE by default.}
}
\value{
Metrics data frame
}
\description{
For each variable it returns: Quantity and percentage of zeros (q_zeros and p_zeros respectevly). Same metrics for NA values (q_NA/p_na), and infinite values (q_inf/p_inf). Last two columns indicates data type and quantity of unique values.
This function print and return the results.
}
\examples{
df_status(heart_disease)
}

