% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecting_best_variables.R
\name{discretize_df}
\alias{discretize_df}
\title{Discretize a data frame}
\usage{
discretize_df(data, data_bins)
}
\arguments{
\item{data}{Input data frame}

\item{data_bins}{data frame generated by 'discretize_get_bins' function. It contains the variable name and the thresholds for each bin, or segment.}
}
\value{
Data frame with the transformed variables
}
\description{
Converts a numerical variable into factor or character, depending on 'stringsAsFactors' parameter.
Binning criteria is equal frequency.
The thresholds for each segment in each variable are generated with 'discretize_get_bins' function, which returns a data frame
that has to be the input in 'data_bins' parameter.
Important to note that the returned data frame contains the non-transformed variables plus the transformed ones.
More info about converting numerical into categorical variables can be found at: <https://livebook.datascienceheroes.com/data-preparation.html#data_types>
}
\examples{
\dontrun{
# Getting the bins thresholds for each. If input is missing, will run for all numerical variables.
d_bins=discretize_get_bins(data=heart_disease, input=c("resting_blood_pressure", "oldpeak"), n_bins=5)

# Now it can be applied on the same data frame, or in a new one (for example in a predictive model that change data over time)
heart_disease_discretized=discretize_df(data=heart_disease, data_bins=d_bins, stringsAsFactors=T)

}
}
