% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecting_best_variables.R
\name{discretize_get_bins}
\alias{discretize_get_bins}
\title{xx}
\usage{
discretize_get_bins(data, input = NULL, n_bins = 5, stringsAsFactors = F)
}
\arguments{
\item{data}{data frame source}

\item{input}{xx}

\item{n_bins}{XX}

\item{stringsAsFactors}{xx}
}
\value{
xxx
}
\description{
xxx
}
\examples{
\dontrun{
# Reducing quantity of countries based on has_flu variable
xxx
}
}
