% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{entropy_2}
\alias{entropy_2}
\title{Computes the entropy between two variables}
\usage{
entropy_2(input, target)
}
\arguments{
\item{x}{input numeric vector}

\item{y}{input numeric vector}
}
\description{
It calculates the entropy between two categorical variables using log2.
This log2 is mentioned in most of the Claude Shannon bibliography.
Input vectors, 'x' and 'y' can be numeric or character.
}
\examples{
\dontrun{
# Measuring entropy between an input and target variable
entropy_2(input=heart_disease$gender, target=heart_disease$has_heart_disease)
}
}
