% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_data_analysis.R
\name{plot_num}
\alias{plot_num}
\title{Plotting numerical data}
\usage{
plot_num(data, bins = 10)
}
\arguments{
\item{data}{data frame}

\item{bins}{number of bars (bins) to plot each histogram, 10 by default}
}
\value{
plot containing all numerical variables
}
\description{
Retrieves one plot containing all the histograms for numerical variables. NA values will not be displayed.
}
\examples{
plot_num(mtcars)
}
