% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numbers.R
\name{prep_outliers}
\alias{prep_outliers}
\title{Remove or treat outliers}
\usage{
prep_outliers(data, str_input, prob)
}
\arguments{
\item{data}{data frame}

\item{str_input}{string input variable (if empty, it runs for all numeric variable), it can take a single character value or a character vector.}

\item{prob}{value from 0 to 1, represents the prob to}
}
\value{
ddd
}
\description{
Treat outliers by removing them or to stopping values at a certain percentile
}
\examples{
prep_outliers()
}

