% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numbers.R
\name{range01}
\alias{range01}
\title{Transform a variable into the 0 to 1 range}
\usage{
range01(x)
}
\arguments{
\item{x}{numeric input vector}
}
\value{
vector ranged into 0-1
}
\description{
Range a variable into [0-1], assigning 0 to the min and 1 to the max of the input variable.
}
\examples{
range01(mtcars$cyl)
}

