% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{range01}
\alias{range01}
\title{Transform a variable into the [0-1] range}
\usage{
range01(var)
}
\arguments{
\item{var}{numeric input vector}
}
\value{
vector with the values scaled into the 0 to 1 range
}
\description{
Range a variable into [0-1], assigning 0 to the min and 1 to the max of the input variable. All NA values will be removed.
}
\examples{
range01(mtcars$cyl)
}

